+++ src/core/rule/configs/rule-no-unknown/rule-no-unknown.config.ts
@@ -1,1 +1,12 @@
-export const RULE_NO_UNKNOWN = ['mixin', 'include', 'extend', 'content', 'each', 'function', 'return', 'if', 'else'];
+export const RULE_NO_UNKNOWN = [
+	'mixin',
+	'include',
+	'extend',
+	'content',
+	'each',
+	'function',
+	'return',
+	'if',
+	'else',
+	'use',
+];